Provide default route for site visitors
User Story:
As a site visitor
I want to visit the site home page and about page
in order to learn about the site and its features

In this commit, I generate a controller named Pages which will be used
to create web pages that are not associated with a model. Additional
pages may be created, but as an initial step, I run the command:

`rails generate controller index about`

This creates the `index.html.erb` and `about.html.erb` files along with
placeholders for customization via scripts, styles and helpers.