Modify About Path to Display at Root
I have created a PagesController to serve pages that do not have an
underlying model. I have created an _about_ view that is related to this
controller.

Before this change, the about page could be retrieved using:
http://localhost:3000/pages/about

I want this page to be accessible from the root:
http://localhost:3000/about

In routes.rb, I changed:

  get 'pages/about'

to

  get 'about' => 'pages#about'

This change also required an edit to the top navigation stored in the
application.html.erb template.