Implement User Roles in the User Model
In this commit, we run the command:

`rails g devise:views`

This applies Bootstrap formatting to our registration and sign-in forms.

Next, we modify the 20151104140144_devise_create_users.rb file and add
a field to the model:

`t.integer :role, default: 0`

Here, we add the Role column, and set the default value to 0. When we
create new users, they will automatically be assigned the role with
ID=0.

In user.rb, we add the following line:

`enum role: [:standard, :premium, :admin]`

This ties our role field in the User model to an enumeration. The first
index of the array is :standard (index = 0). This clarifies that our
default user role is the standard user.

Finally, in our simple_form for Wikis, we replace:

<%= f.input :private %>

With:

<%= f.input :private if current_user.premium? || current_user.admin? %>

Now, you must be a premium user or an administrator in order to create
private wikis.