Implement Basic Stripe Payment Processing for Premium Accounts
In this commit, we first modify the User model to include a name field.

We modify the database.yml file to utilize a Postgresql database instead
of SQLite.

We modify the seeds.rb file to populate the name field when seeding the
database with users.

We implement the ChargesController, a charge processing form, and a Stripe
initializer.

We modify routes.rb so we can direct users to our payment processing
view app/views/charges/new.html.erg.