Add Functionality to Charges Controller for Downgrading Accounts
Modify the charges controller to include `edit` and `destroy` action
methods. In the Edit action, we display thes status of a user
subscription. In the Destroy action, we delete the user's subscription.

The StripeAPI is returning an error when we attempt to remove a
subscription... It states that the user does not have a subscription
(while we can confirm in the Stripe dashboard that the subscription
exists).

This needs further troubleshooting and review before merging into the
master branch.