HOTFIX: Fix Pundit Policy for Premium Users
I have been attempting to get the authorization settings for this
application to function in the following way:

When logged in as a premium user viewing a private wiki that I created,
I can perform a delete operation on the wiki.

When logged in as a premium user and viewing a public wiki,
I cannot perform a delete operation.

I have been struggling with this for the past few days. I implemented
this several ways in the Pundit policy, but nothing that I did appeared
to make much difference. I kept getting the following error:

private method `private' called for #<Class:0x007f84a8857418>

And I THOUGHT this was referring to where I was calling wiki.private in
the Pundit policy file. When I checked the actual server output, it said
more:

ActionView::Template::Error (private method `private' called for
    24:     <button class="btn btn-default">
    25:       <%= link_to 'Edit', edit_wiki_article_path %>
    26:     </button>
    27:     <% if policy(Wiki).destroy? %>
    28:     <button class="btn btn-default">
    29:       <%= link_to "Delete", [@wiki, @article], method: :delete
%>
    30:     </button>
  app/policies/wiki_policy.rb:27:in `destroy?'
  app/views/articles/show.html.erb:27:in
`_app_views_articles_show_html_erb___1939826153190227200_70103881165740'

The word TEMPLATE stood out to me. And the issue was that I was trying
to check the policy on the Wiki class  policy(Wiki) instead of the
instance.

Replacing this line in both the articles and wiki `show.html.erb`
resolved the issue:

<% if policy(Wiki).destroy? %>

changed to

<% if policy(@wiki).destroy? %>