Implement Friendly_IDs for Articles and Wikis
In this commit, we run the command:

  rails generate friendly_id

This creates the database table required to manage slugs for various
models or 'content types.'

Next, we modify the Wiki and Article models in order to take advantage
of this feature. We add a slug field to the wiki and article
migrations, and we edit the models to utilize the slug field. When
either are updated, we update the slug so that it matches the current
title. We also implement the friendly_id history feature so that prior
article and wiki titles are maintained. An article or wiki can be
retrieved using either the current slug or a previous one.